Add auth token requirement and update widget
